fix(feat): handle getting score for different players
The previous test has forced us to actually look at the request's URL
and make a decision. So whilst in our heads, we may have been worrying
about player stores and interfaces the next logical step actually seems
to be about routing.

If we had started with the store code the amount of changes we'd have to
do would be very large compared to this. This is a smaller step towards
our final goal and was driven by tests.

We're resisting the temptation to use any routing libraries right now,
just the smallest step to get our test passing.