rfxr(feat): add interface PlayerStore
It doesn't feel right that our server knows the scores. Our refactoring
has made it pretty clear what to do.

We moved the score calculation out of the main body of our handler into
a function GetPlayerScore. This feels like the right place to separate
the concerns using interfaces.

Let's move our function we re-factored to be an interface instead.