rfxr(feat): implement Handler interface for PlayerServer
Finally, we will now implement the Handler interface by adding a method
to our new struct and putting in our existing handler code.

The only other change is we now call our store.GetPlayerScore to get the
score, rather than the local function we defined (which we can now
delete).