rfxr(test): utilize type PlayerServer
We need to change our tests to instead create a new instance of our
PlayerServer and then call its method ServeHTTP.

Notice we're still not worrying about making stores just yet, we just
want the compiler passing as soon as we can. You should be in the habit
of prioritising having code that compiles and then code that passes the
tests.

By adding more functionality (like stub stores) whilst the code isn't
compiling, we are opening ourselves up to potentially more compilation
problems.