rfxr(test): utilize StubPlayerStore
Now let's create one of these stores for our tests and send it into our
PlayerServer.

Our tests now pass and are looking better. The intent behind our code is
clearer now due to the introduction of the store. We're telling the
reader that because we have this data in a PlayerStore that when you use
it with a PlayerServer you should get the following responses.