fix(cmd): add type InMemoryPlayerStore
Now our tests are passing the last thing we need to do to complete this
refactor is to check if our application is working.

The program should start up but you'll get a horrible response if you
try and hit the server at http://localhost:5000/players/Pepper.

The reason for this is that we have not passed in a PlayerStore.

We'll need to make an implementation of one, but that's difficult right
now as we're not storing any meaningful data so it'll have to be
hard-coded for the time being.