fix(feat): satisfy test for unknown player scenario
Sometimes I heavily roll my eyes when TDD advocates say "make sure you
just write the minimal amount of code to make it pass" as it can feel
very pedantic. But this scenario illustrates the example well. I have
done the bare minimum (knowing it is not correct), which is write a
StatusNotFound on all responses but all our tests are passing!

By doing the bare minimum to make the tests pass it can highlight gaps
in your tests. In our case, we are not asserting that we should be
getting a StatusOK when players do exist in the store.