add(test): integration TestWinsRecordingAndRetrieval()
We are creating our two components we are trying to integrate with,
InMemoryPlayerStore and PlayerServer.

We then fire off 3 requests to record 3 wins for player. We're not too
concerned about the status codes in this test as it's not relevant to
whether they are integrating well.

The next response we do care about (so we store a variable response)
because we are going to try and get the player's score.