rfxr(feat): implement NewPlayerServer()
It's quite odd (and inefficient) to be setting up a router as a request
comes in and then calling it. What we ideally want to do is have some
kind of NewPlayerServer function which will take our dependencies and do
the one-time setup of creating the router. Each request can then just
use that one instance of the router.

- PlayerServer now needs to store a router.
- We have moved the routing creation out of ServeHTTP and into our
NewPlayerServer so this only has to be done once, not per request.
- You will need to update all the test and production code where we used
to do PlayerServer{&store} with NewPlayerServer(&store).