fix(test): satisfy TestFileSystemStore()
There is a flaw in our implementation. First of all, let's remind
ourselves how io.Reader is defined.

type Reader interface {
    Read(p []byte) (n int, err error)
}

With our file, you can imagine it reading through byte by byte until the
end. What happens if you try to Read a second time?

We want this to pass, but if you run the test it doesn't.The problem is
our Reader has reached the end so there is nothing more to read. We need
a way to tell it to go back to the start.

ReadSeeker is another interface in the standard library that can help.