rfxr(test): add helpers createTempFile(), assertScoreEquals()
Our approach is fairly short-sighted for writes. We can't (easily) just
update one "row" of JSON in a file. We'll need to store the whole new
representation of our database on every write.

How do we write? We'd normally use a Writer but we already have our
ReadSeeker. Potentially we could have two dependencies but the standard
library already has an interface for us ReadWriteSeeker which lets us do
all the things we'll need to do with a file.

type FileSystemPlayerStore struct {
    database io.ReadWriteSeeker
}

It's not too surprising that strings.Reader does not implement
ReadWriteSeeker so what do we do?

We have two choices:

- Create a temporary file for each test. *os.File implements
ReadWriteSeeker. The pro of this is it becomes more of an integration
test, we're really reading and writing from the file system so it will
give us a very high level of confidence. The cons are we prefer unit
tests because they are faster and generally simpler. We will also need
to do more work around creating temporary files and then making sure
they're removed after the test.

- We could use a third party library. Mattetti has written a library
filebuffer which implements the interface we need and doesn't touch the
file system.

I don't think there's an especially wrong answer here, but by choosing
to use a third party library I would have to explain dependency
management!

So we will use files instead.Before adding our test we need to make our
other tests compile by replacing the strings.Reader with an os.File.

Let's create some helper functions which will create a temporary file
with some data inside it, and abstract our score tests.