fix(test): satisfy storing wins
You may be asking yourself why I am doing league[i].Wins++ rather than
player.Wins++.

When you range over a slice you are returned the current index of the
loop (in our case i) and a copy of the element at that index. Changing
the Wins value of a copy won't have any effect on the league slice that
we iterate on. For that reason, we need to get the reference to the
actual value by doing league[i] and then changing that value instead.