rfxr(feat): add type tape
There is some more naivety in the way we are dealing with files which
could create a very nasty bug down the line.

When we RecordWin, we Seek back to the start of the file and then write
the new data—but what if the new data was smaller than what was there
before?

In our current case, this is impossible. We never edit or delete scores
so the data can only get bigger. However, it would be irresponsible for
us to leave the code like this; it's not unthinkable that a delete
scenario could come up.

How will we test for this though? What we need to do is first refactor
our code so we separate out the concern of the kind of data we write,
from the writing. We can then test that separately to check it works how
we hope.

We'll create a new type to encapsulate our "when we write we go from the
beginning" functionality.

Notice that we're only implementing Write now, as it encapsulates the
Seek part. This means our FileSystemStore can just have a reference to a
Writer instead.

Finally, we can get the amazing payoff we wanted by removing the Seek
call from RecordWin. Yes, it doesn't feel much, but at least it means if
we do any other kind of writes we can rely on our Write to behave how we
need it to. Plus it will now let us test the potentially problematic
code separately and fix it.